Add cleanup logic to CSdlGenTest and initialize test status variable; reference data updated

+++ Include/imtsdlgencpp/Test/CSdlGenTest.cpp
@@ -172,6 +172,7 @@ void ExecuteTest(
 
 void CSdlGenTest::initTestCase()
 {
+	m_isAllTestsPassed = true;
 	m_tempOutputDir.setPath("SdlGenTest");
 	if (m_tempOutputDir.exists()){
 		QVERIFY(m_tempOutputDir.removeRecursively());
@@ -284,10 +285,18 @@ void CSdlGenTest::PrinterTest()
 }
 
 
+void CSdlGenTest::cleanup() 
+{
+	m_isAllTestsPassed = m_isAllTestsPassed && !QTest::currentTestFailed();
+}
+
+
 void CSdlGenTest::cleanupTestCase()
 {
 	if (m_tempOutputDir.exists()){
-		QVERIFY(m_tempOutputDir.removeRecursively());
+		if (m_isAllTestsPassed) {
+			QVERIFY(m_tempOutputDir.removeRecursively());
+		}
 	}
 }
